Added a autonomous for blueleft and also fixed various auto functions
